:recycle: Move snap_shot::mock module into separate file
