+++ zia/src/context.rs
@@ -14,8 +14,6 @@
 // You should have received a copy of the GNU General Public License
 // along with this program. If not, see <http://www.gnu.org/licenses/>.
 
-#[cfg(test)]
-use crate::concepts::Concept;
 use crate::{
     and_also::AndAlso,
     associativity::Associativity,
