+++ zia/src/context_search_test/basic_comparison.rs
@@ -5,7 +5,7 @@ use crate::{
     context_search::{
         Comparison, ComparisonReason, ContextSearch, ReductionReason,
     },
-    snap_shot::{mock::MockSnapShot, Reader as SnapShotReader},
+    mock_snap_shot::MockSnapShot,
 };
 use maplit::{hashmap, hashset};
 use std::collections::HashMap;
