+++ zia/src/context_search_test/existence_inference_rule.rs
@@ -4,7 +4,7 @@ use crate::{
     context_cache::ContextCache,
     context_delta::ContextDelta,
     context_search::{ContextSearch, ReductionReason},
-    snap_shot::{mock::MockSnapShot, Reader},
+    mock_snap_shot::MockSnapShot,
 };
 use maplit::{hashmap, hashset};
 use std::collections::HashMap;
